census/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.8.6
requests==2.25.1
pyjwt==2.1.0